requirements.txt
streamlit
pandas
reportlab
httpx
python-dotenv
langchain
langchain-groq
langchain-community
faiss-cpu
sentence-transformers
pypdf
pdfplumber
